Rewrite header and body handlers
The previous implementation was incorrect. This gets us a little closer
to correctness.

Request headers are now stored in a table with lowercase keys to achieve
case-insensitive handling. The headers are kept in tables where .list is
a list of the values for headers that list multiple values or repeat,
and .dict is a table of key=value attributes where the values are lists
of the occurences of each attribute, in case attributes are repeated.

The Content-Length header is now sent in responses that specify a body
and is used to read the body of requests. This fixes the issue where
the full request body was not being read.